+++ app/articles/menu.py
@@ -4,7 +4,7 @@ from flask import current_app, json
 
 from app import get_current_locale
 from app.articles import GC_ARTICLES_CACHE_PREFIX, GC_ARTICLES_NAV_CACHE_TTL
-from app.articles.request import get_content
+from app.articles.api import get_content
 from app.extensions import redis_client
